requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.2
requests>=2.28.0

# LLM APIs
openai>=1.0.0
anthropic>=0.8.0

# Required for visualization and data processing
scipy>=1.9.0
tabulate>=0.9.0

# NVIDIA support
# Note: AgentIQ is not included in this requirements file.
# It must be installed separately from https://github.com/NVIDIA/AgentIQ
# Please see the README.md for installation instructions.
# nim-sdk  # Uncomment if using NVIDIA NIM SDK directly
# langchain-nvidia-ai-endpoints>=0.0.1  # Uncomment if using NVIDIA via LangChain

# Development tools
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0

# Streamlit and visualization
streamlit>=1.32.0
plotly>=5.18.0 